First release to google play store

+++ app/database/DAL.js
@@ -1,4 +1,6 @@
 var SQLite = require('react-native-sqlite-storage');
+import {NetInfo} from 'react-native';
+
 function errorCB(err) {
   console.log("SQL Error: " + err);
 }
@@ -52,16 +54,15 @@ export function insertData(data,table){
     
 }
 
-export function getData() {
-    db.transaction((tx) => {
-        tx.executeSql('SELECT * FROM GAME_DATA', [], (tx, results) => {
-            console.log("Query completed");
+export function getData() {db.transaction((tx) => {
+    tx.executeSql('SELECT * FROM GAME_DATA', [], (tx, results) => {
+        console.log("Query completed");
 
-            // Get rows with Web SQL Database spec compliance.
-            let len = results.rows.length;
-            if(len > 0){
+        // Get rows with Web SQL Database spec compliance.
+        let len = results.rows.length;
+        if (len > 0) {
             let gameLogs = new Array();
-            console.log("Query completed>>>>>>>>>"+len);
+            console.log("Query completed>>>>>>>>>" + len);
             for (let i = 0; i < len; i++) {
                 let row = results.rows.item(i);
                 gameLogs.push(row);
@@ -69,29 +70,40 @@ export function getData() {
                 // console.log(`ID: ${row.ID},DEVICE_ID: ${row.DEVICE_ID}, SOUND_PLAYED: ${row.SOUND_PLAYED}, SOUND_SELECTED: ${row.SOUND_SELECTED}, TIMESTAMP: ${row.TIMESTAMP}, STATUS: ${row.STATUS}, SCORE: ${row.SCORE}, LIVES: ${row.LIVES}`);
             }
 
-            fetch('http://104.198.94.126:3000/users/get_sync_data', {
-                    method: 'POST',
-                    headers: {
-                      'Accept': 'application/json',
-                      'Content-Type': 'application/json',
-                    },
-                    body: JSON.stringify(gameLogs)
-                  }).then((response) => response.json())
-                      .then((responseJson) => {
-                        console.log(JSON.stringify(responseJson.returnIds));
-                            db.transaction((tx) => {
-                               tx.executeSql(`DELETE FROM GAME_DATA WHERE ID IN (${responseJson.returnIds})`, [], (tx, results) => {
-                                  console.log("Records deleted successfully");
-                               })
+            NetInfo.fetch().done((reach) => {
+                if (reach != 'NONE') { //check if network is on/off
+                    //console.log('Initial: ' + reach);
+                    NetInfo.isConnected.fetch().then(isConnected => {   //check if there is network conectivity
+                        fetch('http://104.198.94.126:3000/users/get_sync_data', {
+                                method: 'POST',
+                                headers: {
+                                    'Accept': 'application/json',
+                                    'Content-Type': 'application/json',
+                                },
+                                body: JSON.stringify(gameLogs)
+                            }).then((response) => response.json())
+                            .then((responseJson) => {
+                                console.log(JSON.stringify(responseJson.returnIds));
+                                db.transaction((tx) => {
+                                    tx.executeSql(`DELETE FROM GAME_DATA WHERE ID IN (${responseJson.returnIds})`, [], (tx, results) => {
+                                        console.log("Records deleted successfully");
+                                    })
+                                })
                             })
-                      })
-                      .catch((error) => {
-                        console.error(error);
-                      });
-           }else{
+                            .catch((error) => {
+                                console.error(error);
+                            });
+                    }).catch(error => { 
+                      console.error("No Network connectivity::"+error); 
+                    });
+                }else{
+                    console.log('Network is offline');      
+                }
+            });
+        } else {
             console.log('No data to sync');
-           }           
-        });
+        }
     });
+});
 }
